Fix shebang for MacOS/OSX
OSX is shipped with 'bash' version 3.2.57, and this can be found on the '/bin/bash' path. Unfortunately this version doesn't recognise nor the 'readarray' neither the "mapfile" builtins.

Some of the users who actively use bash already installed a newer
version (e.g. from ports or homebrew) but this aren't taken into account since those versions are installed on an other path however are registered correctly.

The best solution would be to use '/usr/bin/env' to pick the configured
version of bash.